Implemented the nesting feature for the Machine and Designed Domains.
If double clicking on these nodes, a problem diagram will be opened. If
the node has no diagram associated, an empty one will be created.

The xtext syntax is updated so that a Node can refer to a subproblem,
then the generated OpenDiagramEditorPolicy is triggered to open the
subproblem diagram instead.

git-svn-id: https://computing-research.open.ac.uk/repos/openpf/trunk@45 650c8b6b-ff4b-431b-9fdd-e90af81125e7