Fix the UNCAL format saved from the xtext and roundtrip back to EMF:
According to Prof. Hu, it is not possible to store information on the
node, such as class name. Instead, it should be added as an edge
"ObjectType". This makes it necessary to preprocess all the nodes when
loading the uncal file because otherwise the abstract class may be
instantiated rather than the concrete class).

Another fix is on the xtext grammar when the following kind of
production rule is to be avoided:
  A: v1= B | v2=C

It is now better to create an abstract class:
  A: BA | CA;
  BA: v=B;
  CA: v=C;
(e.g., A=IntTerm, B=INT, C=Var in the EventCalculus.xtext).

This seems to be fine when generating the EMF model, but not possible
when serialising it back to the xtext file.

Another issue to be solved (not urgent) is that the EMF model and xtext
serialisations can not be both done. That's rather strange, perhaps
because the save_emf already consumes the Resource?

git-svn-id: https://computing-research.open.ac.uk/repos/openpf/trunk@99 650c8b6b-ff4b-431b-9fdd-e90af81125e7