+ solved ticket [3] by inserting two config.ini lines in the pf.product    specifiction (default config.ini does not have them!): {{{ eclipse.p2.profile=epp.package.modeling eclipse.p2.data.area=@config.dir/../p2/ }}}
 + In every generated GMF diagram code, the class <foo>DiagramEditorUtil.java
must be patched (+ new lines, - deleted lines) around line 190:
{{{
+ if (modelResource.getURI().toFileString()==null) {
+ 	model.setName("default");
	modelResource
		.save(uk.ac.open.problem.diagram.part.ProblemDiagramEditorUtil
		.getSaveOptions());
+ }
...
- setCharset(WorkspaceSynchronizer.getFile(modelResource));
- setCharset(WorkspaceSynchronizer.getFile(diagramResource));
}}}

 + Problem.xtext is restoring the non-empty syntax. It is realised that in fact
it is not required to have an initial syntax to match an empty file.  What it
requires is to differentiate the creation of a diagram without the model, or
creation of a diagram after the model. In the latter case, the file does not
exist in the workspace, so the method modelResource.getURI().toFileString()
returns null. Only then we can give a 'default' name to the model, otherwise the name conflict between the room model element and diagram would cause an
inconsistency!

git-svn-id: https://computing-research.open.ac.uk/repos/openpf/trunk@238 650c8b6b-ff4b-431b-9fdd-e90af81125e7